+++ src/components/toDoList/toDoListItem/ToDoListItem.tsx
@@ -1,10 +1,15 @@
 import React, { ChangeEvent, useEffect, useState } from 'react';
-import { Button, ButtonGroup, Flex, Input, ListItem, Text, useDisclosure } from '@chakra-ui/react';
+import { Button, ButtonGroup, Flex, Icon, Input, ListItem, Text, useDisclosure } from '@chakra-ui/react';
+import classNames from 'classnames';
+import { FaTag } from 'react-icons/fa';
+
 import ToDoListItemStatusButton from './toDoListItemStatusButton/ToDoListItemStatusButton';
 import ToDoListItemActionsMenu from './toDoListItemActionsMenu/ToDoListItemActionsMenu';
 import ToDoModal from '../../toDoModal/ToDoModal';
+import LabelSelect from '../../labelSelect/LabelSelect';
+import LabelList from '../../labelList/LabelList';
+import { IToDoListItemProps, TLabels } from '../../../types';
 import { renderStatusElement } from '../../../utils';
-import { IToDoListItemProps } from '../../../types';
 
 const ToDoListItem = ({
   onCancel,
@@ -19,12 +24,18 @@ const ToDoListItem = ({
   const { isOpen, onClose, onOpen } = useDisclosure();
   const [inputValue, setInputValue] = useState<string>('');
   const [actionMenuVisibility, setActionMenuVisibility] = useState<boolean>(false);
+  const [labelMenuVisibility, setLabelMenuVisibility] = useState(false);
+  const [labels, setLabels] = useState<TLabels>(toDoItem.labels);
 
   const handleInput = (event: ChangeEvent<HTMLInputElement>) => setInputValue(event.target.value);
 
   const handleEdit = () => {
     onEdit();
     onOpen();
+  };
+
+  const handleLabelsMenuVisibility = () => {
+    setLabelMenuVisibility(prevState => !prevState);
   };
 
   useEffect(() => {
@@ -35,22 +46,30 @@ const ToDoListItem = ({
     <ListItem justifyContent='space-between'>
       {toDoItem.readOnly ? (
         <Flex
+          alignItems='flex-start'
           justifyContent='space-between'
           width='100%'
           onMouseEnter={() => setActionMenuVisibility(true)}
           onMouseLeave={() => setActionMenuVisibility(false)}>
-          <Flex width='100%'>
+          <Flex flex={3}>
             <ToDoListItemStatusButton completed={toDoItem.completed} onClick={() => onStatusChange(toDoItem)} />
-            <Text as={renderStatusElement(toDoItem.completed)} ml={2} onClick={handleEdit}>
-              {toDoItem.task}
-            </Text>
+            <Flex flexDirection='column'>
+              <Text as={renderStatusElement(toDoItem.completed)} ml={2} onClick={handleEdit}>
+                {toDoItem.task}
+              </Text>
+              <Flex flexWrap='wrap' p={1}>
+                <LabelList labels={toDoItem.labels} />
+              </Flex>
+            </Flex>
           </Flex>
-          <ToDoListItemActionsMenu
-            readOnly={toDoItem.readOnly}
-            visible={actionMenuVisibility}
-            onDelete={() => onDelete(toDoItem.id)}
-            onEdit={() => onInlineEdit(toDoItem.id)}
-          />
+          <Flex flex={1} justifyContent='flex-end'>
+            <ToDoListItemActionsMenu
+              readOnly={toDoItem.readOnly}
+              visible={actionMenuVisibility}
+              onDelete={() => onDelete(toDoItem.id)}
+              onEdit={() => onInlineEdit(toDoItem.id)}
+            />
+          </Flex>
         </Flex>
       ) : (
         <Flex flexDirection='column' width='100%'>
@@ -61,21 +80,54 @@ const ToDoListItem = ({
             onBlur={() => setActionMenuVisibility(false)}
             onChange={handleInput}
             onKeyDown={event => {
-              onKeyboardInput(event, toDoItem, inputValue);
+              if (event.key === 'Escape') {
+                setLabels(toDoItem.labels);
+              }
+              onKeyboardInput(event, toDoItem, inputValue, labels);
               setActionMenuVisibility(false);
             }}
           />
-          <ButtonGroup my={2} size='sm'>
-            <Button colorScheme='teal' onClick={() => onSave(toDoItem, inputValue)}>
-              Save
-            </Button>
-            <Button variant='outline' onClick={() => onCancel(toDoItem.id)}>
-              Cancel
-            </Button>
-          </ButtonGroup>
+          <Flex flexWrap='wrap' p={1}>
+            <LabelList labels={labels} />
+          </Flex>
+          <Flex alignItems='center' justifyContent='space-between'>
+            <Flex>
+              <ButtonGroup my={2} size='sm'>
+                <Button colorScheme='teal' onClick={() => onSave(toDoItem, inputValue, labels)}>
+                  Save
+                </Button>
+                <Button
+                  variant='outline'
+                  onClick={() => {
+                    onCancel(toDoItem.id);
+                    setLabels(toDoItem.labels);
+                  }}>
+                  Cancel
+                </Button>
+              </ButtonGroup>
+            </Flex>
+            <Flex justifyContent='flex-end' position='relative' width='50%'>
+              <Button
+                className={classNames({ 'h-pointer-events-none h-touch-events-none': labelMenuVisibility })}
+                size='sm'
+                onClick={handleLabelsMenuVisibility}>
+                <Icon as={FaTag} boxSize='.9em' color='gray.700' />
+              </Button>
+              <LabelSelect
+                dropdownVisibility={labelMenuVisibility}
+                labels={labels}
+                selectStyles={{
+                  container: { top: '35px' }
+                }}
+                toDoItem={toDoItem}
+                onBlur={() => setLabelMenuVisibility(false)}
+                onSetLabels={setLabels}
+              />
+            </Flex>
+          </Flex>
         </Flex>
       )}
-      <ToDoModal isOpen={isOpen} toDoItem={toDoItem} onClose={onClose} />
+      <ToDoModal isOpen={isOpen} labels={labels} toDoItem={toDoItem} onClose={onClose} onSetLabels={setLabels} />
     </ListItem>
   );
 };