+++ src/store/actions/index.ts
@@ -1,3 +1,4 @@
 export { createToDoItem, removeToDoItem, fetchToDoList, setToDoList, updateToDoItem, fetchToDoItem } from './toDo';
+export { createToDoLabel, fetchToDoLabels, setToDoLabels } from './toDoLabels';
 export { setLoading } from './loading';
 export { setError, showError, setSuccess, showSuccess } from './notifications';