+++ src/store/actions/toDoLabels.ts
@@ -1,0 +1,8 @@
+import { createAction } from '@reduxjs/toolkit';
+import { IToDoLabel } from '../../types';
+
+const prefix = 'TODO_LABELS';
+
+export const fetchToDoLabels = createAction(`${prefix}/fetchToDoLabels`);
+export const setToDoLabels = createAction<IToDoLabel[]>(`${prefix}/setToDoLabels`);
+export const createToDoLabel = createAction<IToDoLabel>(`${prefix}/createToDoLabel`);