+++ src/store/selectors/toDoLabels.ts
@@ -1,0 +1,3 @@
+import { IGlobalState } from '../../types';
+
+export const toDoLabelsStateSelector = (state: IGlobalState) => state.toDo.labels;