+++ src/types/components/toDoModalListItem.ts
@@ -1,6 +1,8 @@
 import { IToDoItem } from '../toDo';
+import { TLabels } from './labelList';
 
 export interface IToDoModalListItemProps {
+  labels: TLabels;
   onReadOnly: (readOnly: boolean) => void;
   onSave: () => void;
   readOnly: boolean;