+++ src/types/index.ts
@@ -1,5 +1,6 @@
 // redux, saga and api
 export type { IInitializationState, IToDoItem } from './toDo';
+export type { IToDoLabel } from './toDoLabels';
 export type { ILoading, ILoadingState, IToDoListLoadingMessages, ILoadingDelays } from './loading';
 export type { INotification, INotificationsState } from './notifications';
 export type { IGlobalState } from './store';
@@ -16,6 +17,7 @@ export type { ILoadingProps } from './components/loading';
 export type { ILoadingTextProps } from './components/loadingText';
 export type { IEmptyState } from './components/emptyState';
 export type { IConfirmationModalProps } from './components/confirmationModal';
+export type { ILabelListProps, TLabels } from './components/labelList';
 
 export { ELoadingDelays, EToDoListLoadingKeys } from './loading';
 export { ENotificationIds } from './notifications';