+++ src/types/toDoLabels.ts
@@ -1,0 +1,7 @@
+import { IToDoActionCallbackProps } from './toDo';
+
+export interface IToDoLabel extends IToDoActionCallbackProps {
+  id: string;
+  label: string;
+  value: string;
+}